Moved reducer function into its own file (CM comment)
